requirements.txt
# LangGraph Vision RAG Agent Requirements
langgraph>=0.2.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langsmith>=0.1.0
chromadb>=0.4.0
cohere>=5.0.0
numpy>=1.24.0
Pillow>=10.0.0
streamlit>=1.32.0
python-dotenv>=1.0.0
PyMuPDF>=1.23.0
tqdm>=4.66.0
requests>=2.31.0
pydantic>=2.0.0
